requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
PyMySQL>=1.1,<2.0
gunicorn>=21.2,<23.0
